revisit host build and dynamically include jni headers based on config, transition for host build, and separate out main
