+++ MODULE.bazel
@@ -19,7 +19,6 @@ bazel_dep(name = "bazel_skylib", version = "1.7.1")
 bazel_dep(name = "rules_pkg", version = "1.0.1")
 bazel_dep(name = "aspect_rules_ts", version = "2.4.2")
 bazel_dep(name = "rules_foreign_cc", version = "0.10.1")
-bazel_dep(name = "rules_jni", version = "0.10.1")
 bazel_dep(name = "googletest", version = "1.14.0")
 
 ####### Node.js version #########
@@ -81,7 +80,6 @@ rules_ts_ext.deps()
 use_repo(rules_ts_ext, "npm_typescript")
 
 # Kotlin
-bazel_dep(name = "rules_java", version = "6.5.2")
 bazel_dep(name = "rules_kotlin", version = "1.9.1")
 single_version_override(
     module_name = "rules_kotlin",
